+++ common-repository/src/main/java/cn/qkl/common/repository/mapper/EvidenceWebMapper.java
@@ -1,21 +1,7 @@
 package cn.qkl.common.repository.mapper;
 
-import static cn.qkl.common.repository.mapper.EvidenceWebDynamicSqlSupport.*;
-import static org.mybatis.dynamic.sql.SqlBuilder.*;
-
 import cn.qkl.common.repository.model.EvidenceWeb;
-import java.util.Collection;
-import java.util.List;
-import java.util.Optional;
-import javax.annotation.Generated;
-import org.apache.ibatis.annotations.DeleteProvider;
-import org.apache.ibatis.annotations.InsertProvider;
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Result;
-import org.apache.ibatis.annotations.ResultMap;
-import org.apache.ibatis.annotations.Results;
-import org.apache.ibatis.annotations.SelectProvider;
-import org.apache.ibatis.annotations.UpdateProvider;
+import org.apache.ibatis.annotations.*;
 import org.apache.ibatis.type.JdbcType;
 import org.mybatis.dynamic.sql.BasicColumn;
 import org.mybatis.dynamic.sql.delete.DeleteDSLCompleter;
@@ -31,6 +17,14 @@ import org.mybatis.dynamic.sql.update.UpdateModel;
 import org.mybatis.dynamic.sql.update.render.UpdateStatementProvider;
 import org.mybatis.dynamic.sql.util.SqlProviderAdapter;
 import org.mybatis.dynamic.sql.util.mybatis3.MyBatis3Utils;
+
+import javax.annotation.Generated;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+
+import static cn.qkl.common.repository.mapper.EvidenceWebDynamicSqlSupport.*;
+import static org.mybatis.dynamic.sql.SqlBuilder.isEqualTo;
 
 @Mapper
 public interface EvidenceWebMapper {
@@ -76,8 +70,8 @@ public interface EvidenceWebMapper {
         @Result(column="web_oss_path", property="webOssPath", jdbcType=JdbcType.VARCHAR),
         @Result(column="cert_oss_path", property="certOssPath", jdbcType=JdbcType.VARCHAR),
         @Result(column="pack_oss_path", property="packOssPath", jdbcType=JdbcType.VARCHAR),
-        @Result(column="evidence_phase", property="evidencePhase", jdbcType=JdbcType.BIGINT),
-        @Result(column="evidence_type", property="evidenceType", jdbcType=JdbcType.BIGINT),
+        @Result(column="evidence_phase", property="evidencePhase", jdbcType=JdbcType.INTEGER),
+        @Result(column="evidence_type", property="evidenceType", jdbcType=JdbcType.INTEGER),
         @Result(column="hash", property="hash", jdbcType=JdbcType.VARCHAR),
         @Result(column="create_time", property="createTime", jdbcType=JdbcType.TIMESTAMP),
         @Result(column="update_time", property="updateTime", jdbcType=JdbcType.TIMESTAMP),
